[PATCH] sn: Fix a couple of compiler warnings
From: Pat Gefre <pfg@sgi.com>

Fix a couple of compiler warnings

+++ arch/ia64/sn/io/xswitch.c
@@ -63,7 +63,7 @@ xswitch_info_vhdl_set(xswitch_info_t xswitch_info,
     if (port > XSWITCH_CENSUS_PORT_MAX)
 	return;
 
-    xswitch_info->vhdl[port] = xwidget;
+    xswitch_info->vhdl[(int)port] = xwidget;
 }
 
 vertex_hdl_t
@@ -73,7 +73,7 @@ xswitch_info_vhdl_get(xswitch_info_t xswitch_info,
     if (port > XSWITCH_CENSUS_PORT_MAX)
 	return GRAPH_VERTEX_NONE;
 
-    return xswitch_info->vhdl[port];
+    return xswitch_info->vhdl[(int)port];
 }
 
 /*
@@ -89,7 +89,7 @@ xswitch_info_master_assignment_set(xswitch_info_t xswitch_info,
     if (port > XSWITCH_CENSUS_PORT_MAX)
 	return;
 
-    xswitch_info->master_vhdl[port] = master_vhdl;
+    xswitch_info->master_vhdl[(int)port] = master_vhdl;
 }
 
 vertex_hdl_t
@@ -99,7 +99,7 @@ xswitch_info_master_assignment_get(xswitch_info_t xswitch_info,
     if (port > XSWITCH_CENSUS_PORT_MAX)
 	return GRAPH_VERTEX_NONE;
 
-    return xswitch_info->master_vhdl[port];
+    return xswitch_info->master_vhdl[(int)port];
 }
 
 void
